moved fp16 union
